requirements.txt
discord.py
supabase
flask
